[DEV] add extra compilation flags & correct warning

+++ etk/RegExp.cpp
@@ -66,7 +66,7 @@ const int64_t etk::constConvertionTableSize = sizeof(etk::constConvertionTable) 
 
 void etk::displayElem(const std::vector<char32_t>& _data, int64_t _start, int64_t _stop) {
 	etk::cout<< ETK_BASH_COLOR_NORMAL;
-	for (int64_t iii=_start; iii<_data.size() && iii<_stop ; iii++) {
+	for (int64_t iii=_start; iii<(int64_t)_data.size() && iii<_stop ; iii++) {
 		switch(_data[iii])
 		{
 			case REGEXP_OPCODE_PTHESE_IN:		etk::cout<<ETK_BASH_COLOR_RED		<< (char*)"(" << ETK_BASH_COLOR_NORMAL;		break;
@@ -125,7 +125,7 @@ char * etk::levelSpace(uint32_t _level) {
 
 
 int64_t etk::getLenOfPTheseElem(const std::vector<char32_t>& _data, int64_t _startPos) {
-	if (_startPos>=_data.size()){
+	if (_startPos>=(int64_t)_data.size()){
 		return 0;
 	}
 	int64_t pos = _startPos;
@@ -136,7 +136,7 @@ int64_t etk::getLenOfPTheseElem(const std::vector<char32_t>& _data, int64_t _sta
 		return 0;
 	}
 	// find size ...
-	while (pos < _data.size() ) {
+	while (pos < (int64_t)_data.size() ) {
 		if(_data[pos] == REGEXP_OPCODE_PTHESE_IN) {
 			// find a sub section : 
 			nbOpen++;
@@ -177,7 +177,7 @@ int64_t etk::getLenOfPThese(const std::vector<char32_t>& _data, int64_t _startPo
 	}
 	pos++;
 	// find size ...
-	while (pos < _data.size() ) {
+	while (pos < (int64_t)_data.size() ) {
 		if(_data[pos]==REGEXP_OPCODE_PTHESE_IN) {
 			// find a sub section : 
 			nbOpen++;
@@ -217,7 +217,7 @@ int64_t etk::getLenOfBracket(const std::vector<char32_t>& _data, int64_t _startP
 	}
 	pos++;
 	// find size ...
-	while (pos < _data.size() ) {
+	while (pos < (int64_t)_data.size() ) {
 		if(_data[pos]==REGEXP_OPCODE_BRACKET_OUT) {
 			// Find the end of the [...]
 			// just return the size inside
@@ -250,7 +250,7 @@ int64_t etk::getLenOfBrace(const std::vector<char32_t>& _data, int64_t _startPos
 	}
 	pos++;
 	// find size ...
-	while (pos < _data.size() ) {
+	while (pos < (int64_t)_data.size() ) {
 		if(_data[pos]==REGEXP_OPCODE_BRACE_OUT) {
 			// Find the end of the [...]
 			// just return the size inside
@@ -275,7 +275,7 @@ int64_t etk::getLenOfBrace(const std::vector<char32_t>& _data, int64_t _startPos
 int64_t etk::getLenOfNormal(const std::vector<char32_t>& _data, int64_t _startPos) {
 	int64_t pos = _startPos;
 	// find size ...
-	while (pos < _data.size() ) {
+	while (pos < (int64_t)_data.size() ) {
 		switch(_data[pos]) {
 			case REGEXP_OPCODE_PTHESE_IN:
 			case REGEXP_OPCODE_PTHESE_OUT:
@@ -329,7 +329,7 @@ bool etk::parseBrace(const std::vector<char32_t>& _data, uint32_t& _min, uint32_
 	int32_t firstElement = 0;
 	int32_t SecondElement = 0;
 	
-	while(k < _data.size()) {
+	while(k < (int64_t)_data.size()) {
 		if (_data[k] == ',') {
 			k++;
 			break;
@@ -345,10 +345,10 @@ bool etk::parseBrace(const std::vector<char32_t>& _data, uint32_t& _min, uint32_
 		}
 		k++;
 	}
-	if (k == _data.size()) {
+	if (k == (int64_t)_data.size()) {
 		SecondElement = firstElement;
 	}
-	while(k < _data.size()) {
+	while(k < (int64_t)_data.size()) {
 		if (_data[k] == ',') {
 			TK_ERROR("Can not find a second , in {} at pos " << k);
 			return false;
